+++ crates/heimwatch-collector/src/focus/linux/dbuslib.rs
@@ -1,0 +1,30 @@
+use anyhow::{Result, anyhow};
+use tokio::sync::mpsc;
+
+/// Check if GNOME D-Bus is accessible.
+pub fn try_gnome_dbus_available() -> Result<()> {
+    // D-Bus availability is checked at runtime in the async spawner.
+    // This is a best-effort check; the actual connection happens later.
+    Ok(())
+}
+
+/// Spawns an async task to listen for GNOME D-Bus window focus signals.
+pub async fn spawn_gnome_dbus_listener(_tx: mpsc::Sender<Option<String>>) -> Result<()> {
+    use zbus::Connection;
+
+    let _conn = Connection::session()
+        .await
+        .map_err(|e| anyhow!("D-Bus session connection failed: {}", e))?;
+
+    // TODO: Complete implementation:
+    // 1. Subscribe to org.gnome.Shell signals or properties
+    // 2. Listen for focus window changes
+    // 3. Extract app_id from the focused window
+    // 4. Send on the channel
+    //
+    // For now, this is a stub that keeps the listener alive.
+    log::debug!("GNOME D-Bus listener started (stub implementation)");
+
+    // Keep the listener alive; will be cancelled on shutdown
+    std::future::pending().await
+}
